IOSS: Test structured decomp > 2Billion cells; scalability fixes

+++ packages/seacas/libraries/ioss/src/cgns/Iocgns_StructuredZoneData.h
@@ -103,7 +103,7 @@ namespace Iocgns {
     // ========================================================================
     // Assume the "work" or computational effort required for a
     // block is proportional to the number of cells.
-    size_t work() const { return m_ordinal[0] * m_ordinal[1] * m_ordinal[2]; }
+    size_t work() const { return (size_t)m_ordinal[0] * m_ordinal[1] * m_ordinal[2]; }
 
     std::pair<StructuredZoneData *, StructuredZoneData *> split(int zone_id, double avg_work,
                                                                 double balance, int rank);
